Migration to Trino
