+++ Gemfile
@@ -3,4 +3,5 @@ gemspec
 
 group :development, :test do
   gem 'tiny-presto', '~> 0.0.7'
+  gem 'presto-client'
 end
