--- lib/presto-client.rb
@@ -1,1 +1,0 @@
-require 'presto/client'
