+++ lib/trino-client.rb
@@ -1,0 +1,1 @@
+require 'trino/client'
