+++ lib/trino/client/model_versions/0.173.rb
@@ -1,5 +1,5 @@
 #
-# Presto client for Ruby
+# Trino client for Ruby
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -13,15 +13,15 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-module Presto::Client::ModelVersions
+module Trino::Client::ModelVersions
 
   ####
-  ## lib/presto/client/model_versions/*.rb is automatically generated using "rake modelgen:all" command.
+  ## lib/trino/client/model_versions/*.rb is automatically generated using "rake modelgen:all" command.
   ## You should not edit this file directly. To modify the class definitions, edit
   ## modelgen/model_versions.rb file and run "rake modelgen:all".
   ##
 
-  module V0_149
+  module V0_173
     class Base < Struct
       class << self
         alias_method :new_struct, :new
@@ -243,7 +243,7 @@ module Presto::Client::ModelVersions
     #
 
     class << AggregationNode =
-        Base.new(:id, :source, :group_by, :aggregations, :functions, :masks, :grouping_sets, :step, :sample_weight, :confidence, :hash_symbol)
+        Base.new(:id, :source, :assignments, :grouping_sets, :step, :hash_symbol, :group_id_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -252,22 +252,17 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["groupBy"],
-          hash["aggregations"],
-          hash["functions"] && Hash[hash["functions"].to_a.map! {|k,v| [k, Signature.decode(v)] }],
-          hash["masks"],
           hash["groupingSets"],
           hash["step"] && hash["step"].downcase.to_sym,
-          hash["sampleWeight"],
-          hash["confidence"],
           hash["hashSymbol"],
+          hash["groupIdSymbol"],
         )
         obj
       end
     end
 
     class << ApplyNode =
-        Base.new(:id, :input, :subquery, :correlation)
+        Base.new(:id, :input, :subquery, :subquery_assignments, :correlation)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -277,7 +272,22 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["input"] && PlanNode.decode(hash["input"]),
           hash["subquery"] && PlanNode.decode(hash["subquery"]),
+          hash["subqueryAssignments"] && Assignments.decode(hash["subqueryAssignments"]),
           hash["correlation"],
+        )
+        obj
+      end
+    end
+
+    class << Assignments =
+        Base.new(:assignments)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["assignments"],
         )
         obj
       end
@@ -291,7 +301,7 @@ module Presto::Client::ModelVersions
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["bufferId"] && TaskId.new(hash["bufferId"]),
+          hash["bufferId"],
           hash["finished"],
           hash["bufferedPages"],
           hash["pagesSent"],
@@ -428,7 +438,7 @@ module Presto::Client::ModelVersions
     end
 
     class << DriverStats =
-        Base.new(:create_time, :start_time, :end_time, :queued_time, :elapsed_time, :memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :raw_input_read_time, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_stats)
+        Base.new(:create_time, :start_time, :end_time, :queued_time, :elapsed_time, :memory_reservation, :peak_memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :raw_input_read_time, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_stats)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -441,6 +451,7 @@ module Presto::Client::ModelVersions
           hash["queuedTime"],
           hash["elapsedTime"],
           hash["memoryReservation"],
+          hash["peakMemoryReservation"],
           hash["systemMemoryReservation"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
@@ -477,7 +488,7 @@ module Presto::Client::ModelVersions
     end
 
     class << ErrorCode =
-        Base.new(:code, :name)
+        Base.new(:code, :name, :type)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -486,6 +497,7 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["code"],
           hash["name"],
+          hash["type"] && hash["type"].downcase.to_sym,
         )
         obj
       end
@@ -597,7 +609,7 @@ module Presto::Client::ModelVersions
     end
 
     class << GroupIdNode =
-        Base.new(:id, :source, :grouping_sets, :identity_mappings, :group_id_symbol)
+        Base.new(:id, :source, :grouping_sets, :grouping_set_mappings, :argument_mappings, :group_id_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -607,7 +619,8 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["groupingSets"],
-          hash["identityMappings"],
+          hash["groupingSetMappings"],
+          hash["argumentMappings"],
           hash["groupIdSymbol"],
         )
         obj
@@ -651,7 +664,7 @@ module Presto::Client::ModelVersions
     end
 
     class << IndexSourceNode =
-        Base.new(:id, :index_handle, :table_handle, :lookup_symbols, :output_symbols, :assignments, :effective_tuple_domain)
+        Base.new(:id, :index_handle, :table_handle, :table_layout, :lookup_symbols, :output_symbols, :assignments, :effective_tuple_domain)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -661,6 +674,7 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["indexHandle"] && IndexHandle.decode(hash["indexHandle"]),
           hash["tableHandle"] && TableHandle.decode(hash["tableHandle"]),
+          hash["tableLayout"] && TableLayoutHandle.decode(hash["tableLayout"]),
           hash["lookupSymbols"],
           hash["outputSymbols"],
           hash["assignments"],
@@ -671,7 +685,7 @@ module Presto::Client::ModelVersions
     end
 
     class << Input =
-        Base.new(:connector_id, :schema, :table, :columns)
+        Base.new(:connector_id, :schema, :table, :connector_info, :columns)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -681,6 +695,7 @@ module Presto::Client::ModelVersions
           hash["connectorId"],
           hash["schema"],
           hash["table"],
+          hash["connectorInfo"],
           hash["columns"] && hash["columns"].map {|h| Column.decode(h) },
         )
         obj
@@ -703,8 +718,25 @@ module Presto::Client::ModelVersions
       end
     end
 
+    class << IntersectNode =
+        Base.new(:id, :sources, :output_to_inputs, :outputs)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["id"],
+          hash["sources"] && hash["sources"].map {|h| PlanNode.decode(h) },
+          hash["outputToInputs"],
+          hash["outputs"],
+        )
+        obj
+      end
+    end
+
     class << JoinNode =
-        Base.new(:id, :type, :left, :right, :criteria, :filter, :left_hash_symbol, :right_hash_symbol)
+        Base.new(:id, :type, :left, :right, :criteria, :output_symbols, :filter, :left_hash_symbol, :right_hash_symbol, :distribution_type)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -716,6 +748,7 @@ module Presto::Client::ModelVersions
           hash["left"] && PlanNode.decode(hash["left"]),
           hash["right"] && PlanNode.decode(hash["right"]),
           hash["criteria"] && hash["criteria"].map {|h| EquiJoinClause.decode(h) },
+          hash["outputSymbols"],
           hash["filter"],
           hash["leftHashSymbol"],
           hash["rightHashSymbol"],
@@ -736,21 +769,6 @@ module Presto::Client::ModelVersions
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["count"],
           hash["partial"],
-        )
-        obj
-      end
-    end
-
-    class << LongVariableConstraint =
-        Base.new(:name, :expression)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["name"],
-          hash["expression"],
         )
         obj
       end
@@ -792,22 +810,25 @@ module Presto::Client::ModelVersions
     end
 
     class << OperatorStats =
-        Base.new(:operator_id, :plan_node_id, :operator_type, :add_input_calls, :add_input_wall, :add_input_cpu, :add_input_user, :input_data_size, :input_positions, :get_output_calls, :get_output_wall, :get_output_cpu, :get_output_user, :output_data_size, :output_positions, :blocked_wall, :finish_calls, :finish_wall, :finish_cpu, :finish_user, :memory_reservation, :system_memory_reservation, :blocked_reason, :info)
+        Base.new(:pipeline_id, :operator_id, :plan_node_id, :operator_type, :total_drivers, :add_input_calls, :add_input_wall, :add_input_cpu, :add_input_user, :input_data_size, :input_positions, :sum_squared_input_positions, :get_output_calls, :get_output_wall, :get_output_cpu, :get_output_user, :output_data_size, :output_positions, :blocked_wall, :finish_calls, :finish_wall, :finish_cpu, :finish_user, :memory_reservation, :system_memory_reservation, :blocked_reason, :info)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
+          hash["pipelineId"],
           hash["operatorId"],
           hash["planNodeId"],
           hash["operatorType"],
+          hash["totalDrivers"],
           hash["addInputCalls"],
           hash["addInputWall"],
           hash["addInputCpu"],
           hash["addInputUser"],
           hash["inputDataSize"],
           hash["inputPositions"],
+          hash["sumSquaredInputPositions"],
           hash["getOutputCalls"],
           hash["getOutputWall"],
           hash["getOutputCpu"],
@@ -822,7 +843,44 @@ module Presto::Client::ModelVersions
           hash["memoryReservation"],
           hash["systemMemoryReservation"],
           hash["blockedReason"] && hash["blockedReason"].downcase.to_sym,
-          hash["info"],
+        )
+        obj
+      end
+    end
+
+    class << Output =
+        Base.new(:connector_id, :schema, :table)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["connectorId"],
+          hash["schema"],
+          hash["table"],
+        )
+        obj
+      end
+    end
+
+    class << OutputBufferInfo =
+        Base.new(:type, :state, :can_add_buffers, :can_add_pages, :total_buffered_bytes, :total_buffered_pages, :total_rows_sent, :total_pages_sent, :buffers)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["type"],
+          hash["state"] && hash["state"].downcase.to_sym,
+          hash["canAddBuffers"],
+          hash["canAddPages"],
+          hash["totalBufferedBytes"],
+          hash["totalBufferedPages"],
+          hash["totalRowsSent"],
+          hash["totalPagesSent"],
+          hash["buffers"] && hash["buffers"].map {|h| BufferInfo.decode(h) },
         )
         obj
       end
@@ -929,13 +987,14 @@ module Presto::Client::ModelVersions
     end
 
     class << PipelineStats =
-        Base.new(:first_start_time, :last_start_time, :last_end_time, :input_pipeline, :output_pipeline, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :completed_drivers, :memory_reservation, :system_memory_reservation, :queued_time, :elapsed_time, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries, :drivers)
+        Base.new(:pipeline_id, :first_start_time, :last_start_time, :last_end_time, :input_pipeline, :output_pipeline, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :completed_drivers, :memory_reservation, :system_memory_reservation, :queued_time, :elapsed_time, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries, :drivers)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
+          hash["pipelineId"],
           hash["firstStartTime"],
           hash["lastStartTime"],
           hash["lastEndTime"],
@@ -999,7 +1058,7 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["assignments"],
+          hash["assignments"] && Assignments.decode(hash["assignments"]),
         )
         obj
       end
@@ -1026,7 +1085,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryInfo =
-        Base.new(:query_id, :session, :state, :memory_pool, :scheduled, :self, :field_names, :query, :query_stats, :set_session_properties, :reset_session_properties, :added_prepared_statements, :deallocated_prepared_statements, :started_transaction_id, :clear_transaction_id, :update_type, :output_stage, :failure_info, :error_code, :inputs)
+        Base.new(:query_id, :session, :state, :memory_pool, :scheduled, :self, :field_names, :query, :query_stats, :set_session_properties, :reset_session_properties, :added_prepared_statements, :deallocated_prepared_statements, :started_transaction_id, :clear_transaction_id, :update_type, :output_stage, :failure_info, :error_code, :inputs, :output, :complete_info, :resource_group_name)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1053,6 +1112,9 @@ module Presto::Client::ModelVersions
           hash["failureInfo"] && FailureInfo.decode(hash["failureInfo"]),
           hash["errorCode"] && ErrorCode.decode(hash["errorCode"]),
           hash["inputs"] && hash["inputs"].map {|h| Input.decode(h) },
+          hash["output"] && Output.decode(hash["output"]),
+          hash["completeInfo"],
+          hash["resourceGroupName"],
         )
         obj
       end
@@ -1082,7 +1144,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryStats =
-        Base.new(:create_time, :execution_start_time, :last_heartbeat, :end_time, :elapsed_time, :queued_time, :analysis_time, :distributed_planning_time, :total_planning_time, :finishing_time, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions)
+        Base.new(:create_time, :execution_start_time, :last_heartbeat, :end_time, :elapsed_time, :queued_time, :analysis_time, :distributed_planning_time, :total_planning_time, :finishing_time, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :scheduled, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1109,6 +1171,7 @@ module Presto::Client::ModelVersions
           hash["cumulativeMemory"],
           hash["totalMemoryReservation"],
           hash["peakMemoryReservation"],
+          hash["scheduled"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
           hash["totalUserTime"],
@@ -1121,6 +1184,7 @@ module Presto::Client::ModelVersions
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
+          hash["operatorSummaries"] && hash["operatorSummaries"].map {|h| OperatorStats.decode(h) },
         )
         obj
       end
@@ -1162,7 +1226,7 @@ module Presto::Client::ModelVersions
     end
 
     class << SampleNode =
-        Base.new(:id, :source, :sample_ratio, :sample_type, :rescaled, :sample_weight_symbol)
+        Base.new(:id, :source, :sample_ratio, :sample_type)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1173,15 +1237,13 @@ module Presto::Client::ModelVersions
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["sampleRatio"],
           hash["sampleType"],
-          hash["rescaled"],
-          hash["sampleWeightSymbol"],
         )
         obj
       end
     end
 
     class << SemiJoinNode =
-        Base.new(:id, :source, :filtering_source, :source_join_symbol, :filtering_source_join_symbol, :semi_join_output, :source_hash_symbol, :filtering_source_hash_symbol)
+        Base.new(:id, :source, :filtering_source, :source_join_symbol, :filtering_source_join_symbol, :semi_join_output, :source_hash_symbol, :filtering_source_hash_symbol, :distribution_type)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1202,7 +1264,7 @@ module Presto::Client::ModelVersions
     end
 
     class << SessionRepresentation =
-        Base.new(:query_id, :transaction_id, :client_transaction_support, :user, :principal, :source, :catalog, :schema, :time_zone_key, :locale, :remote_user_address, :user_agent, :start_time, :system_properties, :catalog_properties, :prepared_statements)
+        Base.new(:query_id, :transaction_id, :client_transaction_support, :user, :principal, :source, :catalog, :schema, :time_zone_key, :locale, :remote_user_address, :user_agent, :client_info, :start_time, :system_properties, :catalog_properties, :prepared_statements)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1221,51 +1283,11 @@ module Presto::Client::ModelVersions
           hash["locale"],
           hash["remoteUserAddress"],
           hash["userAgent"],
+          hash["clientInfo"],
           hash["startTime"],
           hash["systemProperties"],
           hash["catalogProperties"],
           hash["preparedStatements"],
-        )
-        obj
-      end
-    end
-
-    class << SharedBufferInfo =
-        Base.new(:state, :can_add_buffers, :can_add_pages, :total_buffered_bytes, :total_buffered_pages, :total_rows_sent, :total_pages_sent, :buffers)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["state"] && hash["state"].downcase.to_sym,
-          hash["canAddBuffers"],
-          hash["canAddPages"],
-          hash["totalBufferedBytes"],
-          hash["totalBufferedPages"],
-          hash["totalRowsSent"],
-          hash["totalPagesSent"],
-          hash["buffers"] && hash["buffers"].map {|h| BufferInfo.decode(h) },
-        )
-        obj
-      end
-    end
-
-    class << Signature =
-        Base.new(:name, :kind, :type_variable_constraints, :long_variable_constraints, :return_type, :argument_types, :variable_arity)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["name"],
-          hash["kind"] && hash["kind"].downcase.to_sym,
-          hash["typeVariableConstraints"] && hash["typeVariableConstraints"].map {|h| TypeVariableConstraint.decode(h) },
-          hash["longVariableConstraints"] && hash["longVariableConstraints"].map {|h| LongVariableConstraint.decode(h) },
-          hash["returnType"],
-          hash["argumentTypes"],
-          hash["variableArity"],
         )
         obj
       end
@@ -1311,7 +1333,7 @@ module Presto::Client::ModelVersions
     end
 
     class << StageStats =
-        Base.new(:scheduling_complete, :get_split_distribution, :schedule_task_distribution, :add_split_distribution, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions)
+        Base.new(:scheduling_complete, :get_split_distribution, :schedule_task_distribution, :add_split_distribution, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1344,6 +1366,7 @@ module Presto::Client::ModelVersions
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
+          hash["operatorSummaries"] && hash["operatorSummaries"].map {|h| OperatorStats.decode(h) },
         )
         obj
       end
@@ -1445,7 +1468,7 @@ module Presto::Client::ModelVersions
     end
 
     class << TableWriterNode =
-        Base.new(:id, :source, :target, :columns, :column_names, :outputs, :sample_weight_symbol, :partitioning_scheme)
+        Base.new(:id, :source, :target, :columns, :column_names, :outputs, :partitioning_scheme)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1458,7 +1481,6 @@ module Presto::Client::ModelVersions
           hash["columns"],
           hash["columnNames"],
           hash["outputs"],
-          hash["sampleWeightSymbol"],
           hash["partitioningScheme"] && PartitioningScheme.decode(hash["partitioningScheme"]),
         )
         obj
@@ -1466,7 +1488,7 @@ module Presto::Client::ModelVersions
     end
 
     class << TaskInfo =
-        Base.new(:task_status, :last_heartbeat, :output_buffers, :no_more_splits, :stats, :needs_plan)
+        Base.new(:task_status, :last_heartbeat, :output_buffers, :no_more_splits, :stats, :needs_plan, :complete)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1475,10 +1497,11 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["taskStatus"] && TaskStatus.decode(hash["taskStatus"]),
           hash["lastHeartbeat"],
-          hash["outputBuffers"] && SharedBufferInfo.decode(hash["outputBuffers"]),
+          hash["outputBuffers"] && OutputBufferInfo.decode(hash["outputBuffers"]),
           hash["noMoreSplits"],
           hash["stats"] && TaskStats.decode(hash["stats"]),
           hash["needsPlan"],
+          hash["complete"],
         )
         obj
       end
@@ -1568,7 +1591,7 @@ module Presto::Client::ModelVersions
     end
 
     class << TopNRowNumberNode =
-        Base.new(:id, :source, :partition_by, :order_by, :orderings, :row_number_symbol, :max_row_count_per_partition, :partial, :hash_symbol)
+        Base.new(:id, :source, :specification, :row_number_symbol, :max_row_count_per_partition, :partial, :hash_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1577,30 +1600,11 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["partitionBy"],
-          hash["orderBy"],
-          hash["orderings"] && Hash[hash["orderings"].to_a.map! {|k,v| [k, v.downcase.to_sym] }],
+          hash["specification"] && Specification.decode(hash["specification"]),
           hash["rowNumberSymbol"],
           hash["maxRowCountPerPartition"],
           hash["partial"],
           hash["hashSymbol"],
-        )
-        obj
-      end
-    end
-
-    class << TypeVariableConstraint =
-        Base.new(:name, :comparable_required, :orderable_required, :variadic_bound)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["name"],
-          hash["comparableRequired"],
-          hash["orderableRequired"],
-          hash["variadicBound"],
         )
         obj
       end
@@ -1658,7 +1662,7 @@ module Presto::Client::ModelVersions
     end
 
     class << WindowNode =
-        Base.new(:id, :source, :specification, :window_functions, :signatures, :hash_symbol, :pre_partitioned_inputs, :pre_sorted_order_prefix)
+        Base.new(:id, :source, :specification, :window_functions, :hash_symbol, :pre_partitioned_inputs, :pre_sorted_order_prefix)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1668,8 +1672,6 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["specification"] && Specification.decode(hash["specification"]),
-          hash["windowFunctions"],
-          hash["signatures"] && Hash[hash["signatures"].to_a.map! {|k,v| [k, Signature.decode(v)] }],
           hash["hashSymbol"],
           hash["prePartitionedInputs"],
           hash["preSortedOrderPrefix"],
