+++ lib/trino/client/model_versions/0.178.rb
@@ -1,5 +1,5 @@
 #
-# Presto client for Ruby
+# Trino client for Ruby
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -13,15 +13,15 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-module Presto::Client::ModelVersions
+module Trino::Client::ModelVersions
 
   ####
-  ## lib/presto/client/model_versions/*.rb is automatically generated using "rake modelgen:all" command.
+  ## lib/trino/client/model_versions/*.rb is automatically generated using "rake modelgen:all" command.
   ## You should not edit this file directly. To modify the class definitions, edit
   ## modelgen/model_versions.rb file and run "rake modelgen:all".
   ##
 
-  module V316
+  module V0_178
     class Base < Struct
       class << self
         alias_method :new_struct, :new
@@ -66,22 +66,6 @@ module Presto::Client::ModelVersions
       end
 
       attr_reader :query_id, :stage_id, :id
-    end
-
-    class Lifespan < String
-      def initialize(str)
-        super
-        if str == "TaskWide"
-          @grouped = false
-          @group_id = 0
-        else
-          # Group1
-          @grouped = true
-          @group_id = str[5..-1].to_i
-        end
-      end
-
-      attr_reader :grouped, :group_id
     end
 
     class ConnectorSession < Hash
@@ -130,8 +114,6 @@ module Presto::Client::ModelVersions
           when "explainAnalyze"     then ExplainAnalyzeNode
           when "apply"              then ApplyNode
           when "assignUniqueId"     then AssignUniqueId
-          when "lateralJoin"        then LateralJoinNode
-          when "statisticsWriterNode" then StatisticsWriterNode
         end
         if model_class
            node = model_class.decode(hash)
@@ -198,25 +180,9 @@ module Presto::Client::ModelVersions
         end
         obj = allocate
         model_class = case hash["@type"]
-            when "CreateTarget"       then CreateTarget
-            when "InsertTarget"       then InsertTarget
-            when "DeleteTarget"       then DeleteTarget
-        end
-        if model_class
-           model_class.decode(hash)
-        end
-      end
-    end
-
-    class << WriteStatisticsTarget =
-        Base.new(:type, :handle)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        model_class = case hash["@type"]
-            when "WriteStatisticsHandle"       then WriteStatisticsHandle
+            when "CreateHandle"       then CreateHandle
+            when "InsertHandle"       then InsertHandle
+            when "DeleteHandle"       then DeleteHandle
         end
         if model_class
            model_class.decode(hash)
@@ -237,9 +203,6 @@ module Presto::Client::ModelVersions
           when "splitOperator"          then SplitOperatorInfo
           when "hashCollisionsInfo"     then HashCollisionsInfo
           when "partitionedOutput"      then PartitionedOutputInfo
-          when "joinOperatorInfo"       then JoinOperatorInfo
-          when "windowInfo"             then WindowInfo
-          when "tableWriter"            then TableWriterInfo
         end
         if model_class
            model_class.decode(hash)
@@ -263,30 +226,20 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class ResourceGroupId < Array
-      def initialize(array)
-        super()
-        concat(array)
-      end
-    end
-
     ##
     # Those model classes are automatically generated
     #
 
     class << Aggregation =
-        Base.new(:signature, :arguments, :distinct, :filter, :ordering_scheme, :mask)
+        Base.new(:call, :signature, :mask)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
+          hash["call"],
           hash["signature"] && Signature.decode(hash["signature"]),
-          hash["arguments"],
-          hash["distinct"],
-          hash["filter"],
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
           hash["mask"],
         )
         obj
@@ -294,7 +247,7 @@ module Presto::Client::ModelVersions
     end
 
     class << AggregationNode =
-        Base.new(:id, :source, :aggregations, :grouping_sets, :pre_grouped_symbols, :step, :hash_symbol, :group_id_symbol)
+        Base.new(:id, :source, :assignments, :grouping_sets, :step, :hash_symbol, :group_id_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -303,9 +256,8 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["aggregations"] && Hash[hash["aggregations"].to_a.map! {|k,v| [k, Aggregation.decode(v)] }],
-          hash["groupingSets"] && GroupingSetDescriptor.decode(hash["groupingSets"]),
-          hash["preGroupedSymbols"],
+          hash["assignments"] && Hash[hash["assignments"].to_a.map! {|k,v| [k, Aggregation.decode(v)] }],
+          hash["groupingSets"],
           hash["step"] && hash["step"].downcase.to_sym,
           hash["hashSymbol"],
           hash["groupIdSymbol"],
@@ -314,24 +266,8 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << AnalyzeTableHandle =
-        Base.new(:catalog_name, :transaction_handle, :connector_handle)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["catalogName"],
-          hash["transactionHandle"],
-          hash["connectorHandle"],
-        )
-        obj
-      end
-    end
-
     class << ApplyNode =
-        Base.new(:id, :input, :subquery, :subquery_assignments, :correlation, :origin_subquery)
+        Base.new(:id, :input, :subquery, :subquery_assignments, :correlation)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -343,21 +279,20 @@ module Presto::Client::ModelVersions
           hash["subquery"] && PlanNode.decode(hash["subquery"]),
           hash["subqueryAssignments"] && Assignments.decode(hash["subqueryAssignments"]),
           hash["correlation"],
-          hash["originSubquery"],
         )
         obj
       end
     end
 
     class << ArgumentBinding =
-        Base.new(:expression, :constant)
+        Base.new(:column, :constant)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["expression"],
+          hash["column"],
           hash["constant"],
         )
         obj
@@ -389,65 +324,6 @@ module Presto::Client::ModelVersions
         obj = allocate
         obj.send(:initialize_struct,
           hash["assignments"],
-        )
-        obj
-      end
-    end
-
-    class << BasicQueryInfo =
-        Base.new(:query_id, :session, :resource_group_id, :state, :memory_pool, :scheduled, :self, :query, :prepared_query, :query_stats, :error_type, :error_code)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["queryId"],
-          hash["session"] && SessionRepresentation.decode(hash["session"]),
-          hash["resourceGroupId"] && ResourceGroupId.new(hash["resourceGroupId"]),
-          hash["state"] && hash["state"].downcase.to_sym,
-          hash["memoryPool"],
-          hash["scheduled"],
-          hash["self"],
-          hash["query"],
-          hash["preparedQuery"],
-          hash["queryStats"] && BasicQueryStats.decode(hash["queryStats"]),
-          hash["errorType"] && hash["errorType"].downcase.to_sym,
-          hash["errorCode"] && ErrorCode.decode(hash["errorCode"]),
-        )
-        obj
-      end
-    end
-
-    class << BasicQueryStats =
-        Base.new(:create_time, :end_time, :queued_time, :elapsed_time, :execution_time, :total_drivers, :queued_drivers, :running_drivers, :completed_drivers, :raw_input_data_size, :raw_input_positions, :cumulative_user_memory, :user_memory_reservation, :total_memory_reservation, :peak_user_memory_reservation, :peak_total_memory_reservation, :total_cpu_time, :total_scheduled_time, :fully_blocked, :blocked_reasons, :progress_percentage)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["createTime"],
-          hash["endTime"],
-          hash["queuedTime"],
-          hash["elapsedTime"],
-          hash["executionTime"],
-          hash["totalDrivers"],
-          hash["queuedDrivers"],
-          hash["runningDrivers"],
-          hash["completedDrivers"],
-          hash["rawInputDataSize"],
-          hash["rawInputPositions"],
-          hash["cumulativeUserMemory"],
-          hash["userMemoryReservation"],
-          hash["totalMemoryReservation"],
-          hash["peakUserMemoryReservation"],
-          hash["peakTotalMemoryReservation"],
-          hash["totalCpuTime"],
-          hash["totalScheduledTime"],
-          hash["fullyBlocked"],
-          hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["progressPercentage"],
         )
         obj
       end
@@ -488,7 +364,7 @@ module Presto::Client::ModelVersions
     end
 
     class << ClientStageStats =
-        Base.new(:stage_id, :state, :done, :nodes, :total_splits, :queued_splits, :running_splits, :completed_splits, :cpu_time_millis, :wall_time_millis, :processed_rows, :processed_bytes, :sub_stages)
+        Base.new(:stage_id, :state, :done, :nodes, :total_splits, :queued_splits, :running_splits, :completed_splits, :user_time_millis, :cpu_time_millis, :wall_time_millis, :processed_rows, :processed_bytes, :sub_stages)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -503,6 +379,7 @@ module Presto::Client::ModelVersions
           hash["queuedSplits"],
           hash["runningSplits"],
           hash["completedSplits"],
+          hash["userTimeMillis"],
           hash["cpuTimeMillis"],
           hash["wallTimeMillis"],
           hash["processedRows"],
@@ -545,21 +422,6 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << Code =
-        Base.new(:code, :name)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["code"],
-          hash["name"],
-        )
-        obj
-      end
-    end
-
     class << Column =
         Base.new(:name, :type)
       def decode(hash)
@@ -575,22 +437,7 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << ColumnStatisticMetadata =
-        Base.new(:column_name, :statistic_type)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["columnName"],
-          hash["statisticType"] && hash["statisticType"].downcase.to_sym,
-        )
-        obj
-      end
-    end
-
-    class << CreateTarget =
+    class << CreateHandle =
         Base.new(:handle, :schema_table_name)
       def decode(hash)
         unless hash.is_a?(Hash)
@@ -599,6 +446,21 @@ module Presto::Client::ModelVersions
         obj = allocate
         obj.send(:initialize_struct,
           hash["handle"] && OutputTableHandle.decode(hash["handle"]),
+          hash["schemaTableName"] && SchemaTableName.decode(hash["schemaTableName"]),
+        )
+        obj
+      end
+    end
+
+    class << DeleteHandle =
+        Base.new(:handle, :schema_table_name)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["handle"] && TableHandle.decode(hash["handle"]),
           hash["schemaTableName"] && SchemaTableName.decode(hash["schemaTableName"]),
         )
         obj
@@ -615,24 +477,9 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["target"] && DeleteTarget.decode(hash["target"]),
+          hash["target"] && DeleteHandle.decode(hash["target"]),
           hash["rowId"],
           hash["outputs"],
-        )
-        obj
-      end
-    end
-
-    class << DeleteTarget =
-        Base.new(:handle, :schema_table_name)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["handle"] && TableHandle.decode(hash["handle"]),
-          hash["schemaTableName"] && SchemaTableName.decode(hash["schemaTableName"]),
         )
         obj
       end
@@ -658,33 +505,27 @@ module Presto::Client::ModelVersions
     end
 
     class << DriverStats =
-        Base.new(:lifespan, :create_time, :start_time, :end_time, :queued_time, :elapsed_time, :user_memory_reservation, :revocable_memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :physical_input_data_size, :physical_input_positions, :physical_input_read_time, :internal_network_input_data_size, :internal_network_input_positions, :internal_network_input_read_time, :raw_input_data_size, :raw_input_positions, :raw_input_read_time, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :physical_written_data_size, :operator_stats)
+        Base.new(:create_time, :start_time, :end_time, :queued_time, :elapsed_time, :memory_reservation, :peak_memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :raw_input_read_time, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_stats)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["lifespan"] && Lifespan.new(hash["lifespan"]),
           hash["createTime"],
           hash["startTime"],
           hash["endTime"],
           hash["queuedTime"],
           hash["elapsedTime"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["memoryReservation"],
+          hash["peakMemoryReservation"],
           hash["systemMemoryReservation"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
+          hash["totalUserTime"],
           hash["totalBlockedTime"],
           hash["fullyBlocked"],
           hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["physicalInputReadTime"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
-          hash["internalNetworkInputReadTime"],
           hash["rawInputDataSize"],
           hash["rawInputPositions"],
           hash["rawInputReadTime"],
@@ -692,27 +533,7 @@ module Presto::Client::ModelVersions
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
           hash["operatorStats"] && hash["operatorStats"].map {|h| OperatorStats.decode(h) },
-        )
-        obj
-      end
-    end
-
-    class << DriverWindowInfo =
-        Base.new(:sum_squared_differences_positions_of_index, :sum_squared_differences_size_of_index, :sum_squared_differences_size_in_partition, :total_partitions_count, :total_rows_count, :number_of_indexes)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["sumSquaredDifferencesPositionsOfIndex"],
-          hash["sumSquaredDifferencesSizeOfIndex"],
-          hash["sumSquaredDifferencesSizeInPartition"],
-          hash["totalPartitionsCount"],
-          hash["totalRowsCount"],
-          hash["numberOfIndexes"],
         )
         obj
       end
@@ -785,7 +606,7 @@ module Presto::Client::ModelVersions
     end
 
     class << ExchangeNode =
-        Base.new(:id, :type, :scope, :partitioning_scheme, :sources, :inputs, :ordering_scheme)
+        Base.new(:id, :type, :scope, :partitioning_scheme, :sources, :inputs)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -798,14 +619,13 @@ module Presto::Client::ModelVersions
           hash["partitioningScheme"] && PartitioningScheme.decode(hash["partitioningScheme"]),
           hash["sources"] && hash["sources"].map {|h| PlanNode.decode(h) },
           hash["inputs"],
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
         )
         obj
       end
     end
 
     class << ExecutionFailureInfo =
-        Base.new(:type, :message, :cause, :suppressed, :stack, :error_location, :error_code, :semantic_error_code, :remote_host)
+        Base.new(:type, :message, :cause, :suppressed, :stack, :error_location, :error_code, :remote_host)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -819,7 +639,6 @@ module Presto::Client::ModelVersions
           hash["stack"],
           hash["errorLocation"] && ErrorLocation.decode(hash["errorLocation"]),
           hash["errorCode"] && ErrorCode.decode(hash["errorCode"]),
-          hash["semanticErrorCode"] && hash["semanticErrorCode"].downcase.to_sym,
           hash["remoteHost"],
         )
         obj
@@ -827,7 +646,7 @@ module Presto::Client::ModelVersions
     end
 
     class << ExplainAnalyzeNode =
-        Base.new(:id, :source, :output_symbol, :verbose)
+        Base.new(:id, :source, :output_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -837,7 +656,6 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["outputSymbol"],
-          hash["verbose"],
         )
         obj
       end
@@ -879,15 +697,15 @@ module Presto::Client::ModelVersions
     end
 
     class << Function =
-        Base.new(:signature, :arguments, :frame)
+        Base.new(:function_call, :signature, :frame)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
+          hash["functionCall"],
           hash["signature"] && Signature.decode(hash["signature"]),
-          hash["arguments"],
           hash["frame"],
         )
         obj
@@ -895,7 +713,7 @@ module Presto::Client::ModelVersions
     end
 
     class << GroupIdNode =
-        Base.new(:id, :source, :grouping_sets, :grouping_columns, :aggregation_arguments, :group_id_symbol)
+        Base.new(:id, :source, :grouping_sets, :grouping_set_mappings, :argument_mappings, :group_id_symbol)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -905,39 +723,23 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["groupingSets"],
-          hash["groupingColumns"],
-          hash["aggregationArguments"],
+          hash["groupingSetMappings"],
+          hash["argumentMappings"],
           hash["groupIdSymbol"],
         )
         obj
       end
     end
 
-    class << GroupingSetDescriptor =
-        Base.new(:grouping_keys, :grouping_set_count, :global_grouping_sets)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["groupingKeys"],
-          hash["groupingSetCount"],
-          hash["globalGroupingSets"],
-        )
-        obj
-      end
-    end
-
     class << IndexHandle =
-        Base.new(:catalog_name, :transaction_handle, :connector_handle)
+        Base.new(:connector_id, :transaction_handle, :connector_handle)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["catalogName"],
+          hash["connectorId"],
           hash["transactionHandle"],
           hash["connectorHandle"],
         )
@@ -966,7 +768,7 @@ module Presto::Client::ModelVersions
     end
 
     class << IndexSourceNode =
-        Base.new(:id, :index_handle, :table_handle, :lookup_symbols, :output_symbols, :assignments, :current_constraint)
+        Base.new(:id, :index_handle, :table_handle, :table_layout, :lookup_symbols, :output_symbols, :assignments, :effective_tuple_domain)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -976,24 +778,25 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["indexHandle"] && IndexHandle.decode(hash["indexHandle"]),
           hash["tableHandle"] && TableHandle.decode(hash["tableHandle"]),
+          hash["tableLayout"] && TableLayoutHandle.decode(hash["tableLayout"]),
           hash["lookupSymbols"],
           hash["outputSymbols"],
           hash["assignments"],
-          hash["currentConstraint"],
+          hash["effectiveTupleDomain"],
         )
         obj
       end
     end
 
     class << Input =
-        Base.new(:catalog_name, :schema, :table, :connector_info, :columns)
+        Base.new(:connector_id, :schema, :table, :connector_info, :columns)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["catalogName"],
+          hash["connectorId"],
           hash["schema"],
           hash["table"],
           hash["connectorInfo"],
@@ -1003,23 +806,7 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << InsertTableHandle =
-        Base.new(:catalog_name, :transaction_handle, :connector_handle)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["catalogName"],
-          hash["transactionHandle"],
-          hash["connectorHandle"],
-        )
-        obj
-      end
-    end
-
-    class << InsertTarget =
+    class << InsertHandle =
         Base.new(:handle, :schema_table_name)
       def decode(hash)
         unless hash.is_a?(Hash)
@@ -1029,6 +816,22 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["handle"] && InsertTableHandle.decode(hash["handle"]),
           hash["schemaTableName"] && SchemaTableName.decode(hash["schemaTableName"]),
+        )
+        obj
+      end
+    end
+
+    class << InsertTableHandle =
+        Base.new(:connector_id, :transaction_handle, :connector_handle)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
+          hash["connectorId"],
+          hash["transactionHandle"],
+          hash["connectorHandle"],
         )
         obj
       end
@@ -1052,7 +855,7 @@ module Presto::Client::ModelVersions
     end
 
     class << JoinNode =
-        Base.new(:id, :type, :left, :right, :criteria, :output_symbols, :filter, :left_hash_symbol, :right_hash_symbol, :distribution_type, :spillable, :dynamic_filters)
+        Base.new(:id, :type, :left, :right, :criteria, :output_symbols, :filter, :left_hash_symbol, :right_hash_symbol, :distribution_type)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1069,52 +872,13 @@ module Presto::Client::ModelVersions
           hash["leftHashSymbol"],
           hash["rightHashSymbol"],
           hash["distributionType"] && hash["distributionType"].downcase.to_sym,
-          hash["spillable"],
-          hash["dynamicFilters"],
-        )
-        obj
-      end
-    end
-
-    class << JoinOperatorInfo =
-        Base.new(:join_type, :log_histogram_probes, :log_histogram_output, :lookup_source_positions)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["joinType"] && hash["joinType"].downcase.to_sym,
-          hash["logHistogramProbes"],
-          hash["logHistogramOutput"],
-          hash["lookupSourcePositions"],
-        )
-        obj
-      end
-    end
-
-    class << LateralJoinNode =
-        Base.new(:id, :input, :subquery, :correlation, :type, :filter, :origin_subquery)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["id"],
-          hash["input"] && PlanNode.decode(hash["input"]),
-          hash["subquery"] && PlanNode.decode(hash["subquery"]),
-          hash["correlation"],
-          hash["type"],
-          hash["filter"],
-          hash["originSubquery"],
         )
         obj
       end
     end
 
     class << LimitNode =
-        Base.new(:id, :source, :count, :ties_resolving_scheme, :partial)
+        Base.new(:id, :source, :count, :partial)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1124,24 +888,7 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["count"],
-          hash["tiesResolvingScheme"] && OrderingScheme.decode(hash["tiesResolvingScheme"]),
           hash["partial"],
-        )
-        obj
-      end
-    end
-
-    class << LocalCostEstimate =
-        Base.new(:cpu_cost, :max_memory, :network_cost)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["cpuCost"],
-          hash["maxMemory"],
-          hash["networkCost"],
         )
         obj
       end
@@ -1195,15 +942,31 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << OperatorStats =
-        Base.new(:stage_id, :pipeline_id, :operator_id, :plan_node_id, :operator_type, :total_drivers, :add_input_calls, :add_input_wall, :add_input_cpu, :physical_input_data_size, :physical_input_positions, :internal_network_input_data_size, :internal_network_input_positions, :raw_input_data_size, :input_data_size, :input_positions, :sum_squared_input_positions, :get_output_calls, :get_output_wall, :get_output_cpu, :output_data_size, :output_positions, :physical_written_data_size, :blocked_wall, :finish_calls, :finish_wall, :finish_cpu, :user_memory_reservation, :revocable_memory_reservation, :system_memory_reservation, :peak_user_memory_reservation, :peak_system_memory_reservation, :peak_revocable_memory_reservation, :peak_total_memory_reservation, :spilled_data_size, :blocked_reason, :info)
+    class << MetadataDeleteNode =
+        Base.new(:id, :target, :output, :table_layout)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["stageId"],
+          hash["id"],
+          hash["target"] && DeleteHandle.decode(hash["target"]),
+          hash["output"],
+          hash["tableLayout"] && TableLayoutHandle.decode(hash["tableLayout"]),
+        )
+        obj
+      end
+    end
+
+    class << OperatorStats =
+        Base.new(:pipeline_id, :operator_id, :plan_node_id, :operator_type, :total_drivers, :add_input_calls, :add_input_wall, :add_input_cpu, :add_input_user, :input_data_size, :input_positions, :sum_squared_input_positions, :get_output_calls, :get_output_wall, :get_output_cpu, :get_output_user, :output_data_size, :output_positions, :blocked_wall, :finish_calls, :finish_wall, :finish_cpu, :finish_user, :memory_reservation, :system_memory_reservation, :blocked_reason, :info)
+      def decode(hash)
+        unless hash.is_a?(Hash)
+          raise TypeError, "Can't convert #{hash.class} to Hash"
+        end
+        obj = allocate
+        obj.send(:initialize_struct,
           hash["pipelineId"],
           hash["operatorId"],
           hash["planNodeId"],
@@ -1212,32 +975,23 @@ module Presto::Client::ModelVersions
           hash["addInputCalls"],
           hash["addInputWall"],
           hash["addInputCpu"],
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
-          hash["rawInputDataSize"],
+          hash["addInputUser"],
           hash["inputDataSize"],
           hash["inputPositions"],
           hash["sumSquaredInputPositions"],
           hash["getOutputCalls"],
           hash["getOutputWall"],
           hash["getOutputCpu"],
+          hash["getOutputUser"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
           hash["blockedWall"],
           hash["finishCalls"],
           hash["finishWall"],
           hash["finishCpu"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["finishUser"],
+          hash["memoryReservation"],
           hash["systemMemoryReservation"],
-          hash["peakUserMemoryReservation"],
-          hash["peakSystemMemoryReservation"],
-          hash["peakRevocableMemoryReservation"],
-          hash["peakTotalMemoryReservation"],
-          hash["spilledDataSize"],
           hash["blockedReason"] && hash["blockedReason"].downcase.to_sym,
           hash["info"] && OperatorInfo.decode(hash["info"]),
         )
@@ -1245,30 +999,15 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << OrderingScheme =
-        Base.new(:order_by, :orderings)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["orderBy"],
-          hash["orderings"] && Hash[hash["orderings"].to_a.map! {|k,v| [k, v.downcase.to_sym] }],
-        )
-        obj
-      end
-    end
-
     class << Output =
-        Base.new(:catalog_name, :schema, :table)
+        Base.new(:connector_id, :schema, :table)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["catalogName"],
+          hash["connectorId"],
           hash["schema"],
           hash["table"],
         )
@@ -1316,14 +1055,14 @@ module Presto::Client::ModelVersions
     end
 
     class << OutputTableHandle =
-        Base.new(:catalog_name, :transaction_handle, :connector_handle)
+        Base.new(:connector_id, :transaction_handle, :connector_handle)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["catalogName"],
+          hash["connectorId"],
           hash["transactionHandle"],
           hash["connectorHandle"],
         )
@@ -1368,22 +1107,6 @@ module Presto::Client::ModelVersions
           hash["bufferedBytes"],
           hash["rowsAdded"],
           hash["pagesAdded"],
-        )
-        obj
-      end
-    end
-
-    class << PartitionedOutputInfo =
-        Base.new(:rows_added, :pages_added, :output_buffer_peak_memory_usage)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["rowsAdded"],
-          hash["pagesAdded"],
-          hash["outputBufferPeakMemoryUsage"],
         )
         obj
       end
@@ -1439,7 +1162,7 @@ module Presto::Client::ModelVersions
     end
 
     class << PipelineStats =
-        Base.new(:pipeline_id, :first_start_time, :last_start_time, :last_end_time, :input_pipeline, :output_pipeline, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :blocked_drivers, :completed_drivers, :user_memory_reservation, :revocable_memory_reservation, :system_memory_reservation, :queued_time, :elapsed_time, :total_scheduled_time, :total_cpu_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :physical_input_data_size, :physical_input_positions, :internal_network_input_data_size, :internal_network_input_positions, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :physical_written_data_size, :operator_summaries, :drivers)
+        Base.new(:pipeline_id, :first_start_time, :last_start_time, :last_end_time, :input_pipeline, :output_pipeline, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :blocked_drivers, :completed_drivers, :memory_reservation, :system_memory_reservation, :queued_time, :elapsed_time, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries, :drivers)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1459,27 +1182,22 @@ module Presto::Client::ModelVersions
           hash["runningPartitionedDrivers"],
           hash["blockedDrivers"],
           hash["completedDrivers"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["memoryReservation"],
           hash["systemMemoryReservation"],
           hash["queuedTime"] && DistributionSnapshot.decode(hash["queuedTime"]),
           hash["elapsedTime"] && DistributionSnapshot.decode(hash["elapsedTime"]),
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
+          hash["totalUserTime"],
           hash["totalBlockedTime"],
           hash["fullyBlocked"],
           hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
           hash["rawInputDataSize"],
           hash["rawInputPositions"],
           hash["processedInputDataSize"],
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
           hash["operatorSummaries"] && hash["operatorSummaries"].map {|h| OperatorStats.decode(h) },
           hash["drivers"] && hash["drivers"].map {|h| DriverStats.decode(h) },
         )
@@ -1487,26 +1205,8 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << PlanCostEstimate =
-        Base.new(:cpu_cost, :max_memory, :max_memory_when_outputting, :network_cost, :root_node_local_cost_estimate)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["cpuCost"],
-          hash["maxMemory"],
-          hash["maxMemoryWhenOutputting"],
-          hash["networkCost"],
-          hash["rootNodeLocalCostEstimate"] && LocalCostEstimate.decode(hash["rootNodeLocalCostEstimate"]),
-        )
-        obj
-      end
-    end
-
     class << PlanFragment =
-        Base.new(:id, :root, :symbols, :partitioning, :partitioned_sources, :partitioning_scheme, :stage_execution_descriptor, :stats_and_costs, :json_representation)
+        Base.new(:id, :root, :symbols, :partitioning, :partitioned_sources, :partitioning_scheme)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1519,39 +1219,6 @@ module Presto::Client::ModelVersions
           hash["partitioning"] && PartitioningHandle.decode(hash["partitioning"]),
           hash["partitionedSources"],
           hash["partitioningScheme"] && PartitioningScheme.decode(hash["partitioningScheme"]),
-          hash["stageExecutionDescriptor"] && StageExecutionDescriptor.decode(hash["stageExecutionDescriptor"]),
-          hash["statsAndCosts"] && StatsAndCosts.decode(hash["statsAndCosts"]),
-          hash["jsonRepresentation"],
-        )
-        obj
-      end
-    end
-
-    class << PlanNodeStatsEstimate =
-        Base.new(:output_row_count, :symbol_statistics)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["outputRowCount"],
-          hash["symbolStatistics"] && Hash[hash["symbolStatistics"].to_a.map! {|k,v| [k, SymbolStatsEstimate.decode(v)] }],
-        )
-        obj
-      end
-    end
-
-    class << PrestoWarning =
-        Base.new(:warning_code, :message)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["warningCode"] && WarningCode.decode(hash["warningCode"]),
-          hash["message"],
         )
         obj
       end
@@ -1574,7 +1241,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryError =
-        Base.new(:message, :sql_state, :error_code, :error_name, :semantic_error_name, :error_type, :error_location, :failure_info)
+        Base.new(:message, :sql_state, :error_code, :error_name, :error_type, :error_location, :failure_info)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1585,7 +1252,6 @@ module Presto::Client::ModelVersions
           hash["sqlState"],
           hash["errorCode"],
           hash["errorName"],
-          hash["semanticErrorName"],
           hash["errorType"],
           hash["errorLocation"] && ErrorLocation.decode(hash["errorLocation"]),
           hash["failureInfo"] && FailureInfo.decode(hash["failureInfo"]),
@@ -1595,7 +1261,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryInfo =
-        Base.new(:query_id, :session, :state, :memory_pool, :scheduled, :self, :field_names, :query, :prepared_query, :query_stats, :set_catalog, :set_schema, :set_path, :set_session_properties, :reset_session_properties, :set_roles, :added_prepared_statements, :deallocated_prepared_statements, :started_transaction_id, :clear_transaction_id, :update_type, :output_stage, :failure_info, :error_code, :warnings, :inputs, :output, :complete_info, :resource_group_id, :final_query_info)
+        Base.new(:query_id, :session, :state, :memory_pool, :scheduled, :self, :field_names, :query, :query_stats, :set_session_properties, :reset_session_properties, :added_prepared_statements, :deallocated_prepared_statements, :started_transaction_id, :clear_transaction_id, :update_type, :output_stage, :failure_info, :error_code, :inputs, :output, :complete_info, :resource_group_name, :final_query_info)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1610,27 +1276,21 @@ module Presto::Client::ModelVersions
           hash["self"],
           hash["fieldNames"],
           hash["query"],
-          hash["preparedQuery"],
           hash["queryStats"] && QueryStats.decode(hash["queryStats"]),
-          hash["setCatalog"],
-          hash["setSchema"],
-          hash["setPath"],
           hash["setSessionProperties"],
           hash["resetSessionProperties"],
-          hash["setRoles"] && Hash[hash["setRoles"].to_a.map! {|k,v| [k, SelectedRole.decode(v)] }],
           hash["addedPreparedStatements"],
           hash["deallocatedPreparedStatements"],
           hash["startedTransactionId"],
           hash["clearTransactionId"],
           hash["updateType"],
           hash["outputStage"] && StageInfo.decode(hash["outputStage"]),
-          hash["failureInfo"] && ExecutionFailureInfo.decode(hash["failureInfo"]),
+          hash["failureInfo"] && FailureInfo.decode(hash["failureInfo"]),
           hash["errorCode"] && ErrorCode.decode(hash["errorCode"]),
-          hash["warnings"] && hash["warnings"].map {|h| PrestoWarning.decode(h) },
           hash["inputs"] && hash["inputs"].map {|h| Input.decode(h) },
           hash["output"] && Output.decode(hash["output"]),
           hash["completeInfo"],
-          hash["resourceGroupId"] && ResourceGroupId.new(hash["resourceGroupId"]),
+          hash["resourceGroupName"],
           hash["finalQueryInfo"],
         )
         obj
@@ -1638,7 +1298,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryResults =
-        Base.new(:id, :info_uri, :partial_cancel_uri, :next_uri, :columns, :data, :stats, :error, :warnings, :update_type, :update_count)
+        Base.new(:id, :info_uri, :partial_cancel_uri, :next_uri, :columns, :data, :stats, :error, :update_type, :update_count)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1653,7 +1313,6 @@ module Presto::Client::ModelVersions
           hash["data"],
           hash["stats"] && StatementStats.decode(hash["stats"]),
           hash["error"] && QueryError.decode(hash["error"]),
-          hash["warnings"] && hash["warnings"].map {|h| Warning.decode(h) },
           hash["updateType"],
           hash["updateCount"],
         )
@@ -1662,7 +1321,7 @@ module Presto::Client::ModelVersions
     end
 
     class << QueryStats =
-        Base.new(:create_time, :execution_start_time, :last_heartbeat, :end_time, :elapsed_time, :queued_time, :resource_waiting_time, :dispatching_time, :execution_time, :analysis_time, :distributed_planning_time, :total_planning_time, :finishing_time, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :blocked_drivers, :completed_drivers, :cumulative_user_memory, :user_memory_reservation, :revocable_memory_reservation, :total_memory_reservation, :peak_user_memory_reservation, :peak_revocable_memory_reservation, :peak_total_memory_reservation, :peak_task_user_memory, :peak_task_revocable_memory, :peak_task_total_memory, :scheduled, :total_scheduled_time, :total_cpu_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :physical_input_data_size, :physical_input_positions, :internal_network_input_data_size, :internal_network_input_positions, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :physical_written_data_size, :stage_gc_statistics, :operator_summaries)
+        Base.new(:create_time, :execution_start_time, :last_heartbeat, :end_time, :elapsed_time, :queued_time, :analysis_time, :distributed_planning_time, :total_planning_time, :finishing_time, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :blocked_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :scheduled, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :operator_summaries)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1675,9 +1334,6 @@ module Presto::Client::ModelVersions
           hash["endTime"],
           hash["elapsedTime"],
           hash["queuedTime"],
-          hash["resourceWaitingTime"],
-          hash["dispatchingTime"],
-          hash["executionTime"],
           hash["analysisTime"],
           hash["distributedPlanningTime"],
           hash["totalPlanningTime"],
@@ -1690,34 +1346,22 @@ module Presto::Client::ModelVersions
           hash["runningDrivers"],
           hash["blockedDrivers"],
           hash["completedDrivers"],
-          hash["cumulativeUserMemory"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["cumulativeMemory"],
           hash["totalMemoryReservation"],
-          hash["peakUserMemoryReservation"],
-          hash["peakRevocableMemoryReservation"],
-          hash["peakTotalMemoryReservation"],
-          hash["peakTaskUserMemory"],
-          hash["peakTaskRevocableMemory"],
-          hash["peakTaskTotalMemory"],
+          hash["peakMemoryReservation"],
           hash["scheduled"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
+          hash["totalUserTime"],
           hash["totalBlockedTime"],
           hash["fullyBlocked"],
           hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
           hash["rawInputDataSize"],
           hash["rawInputPositions"],
           hash["processedInputDataSize"],
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
-          hash["stageGcStatistics"] && hash["stageGcStatistics"].map {|h| StageGcStatistics.decode(h) },
           hash["operatorSummaries"] && hash["operatorSummaries"].map {|h| OperatorStats.decode(h) },
         )
         obj
@@ -1725,7 +1369,7 @@ module Presto::Client::ModelVersions
     end
 
     class << RemoteSourceNode =
-        Base.new(:id, :source_fragment_ids, :outputs, :ordering_scheme, :exchange_type)
+        Base.new(:id, :source_fragment_ids, :outputs)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1735,24 +1379,6 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["sourceFragmentIds"],
           hash["outputs"],
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
-          hash["exchangeType"] && hash["exchangeType"].downcase.to_sym,
-        )
-        obj
-      end
-    end
-
-    class << ResourceEstimates =
-        Base.new(:execution_time, :cpu_time, :peak_memory)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["executionTime"],
-          hash["cpuTime"],
-          hash["peakMemory"],
         )
         obj
       end
@@ -1809,21 +1435,6 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << SelectedRole =
-        Base.new(:type, :role)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["type"],
-          hash["role"],
-        )
-        obj
-      end
-    end
-
     class << SemiJoinNode =
         Base.new(:id, :source, :filtering_source, :source_join_symbol, :filtering_source_join_symbol, :semi_join_output, :source_hash_symbol, :filtering_source_hash_symbol, :distribution_type)
       def decode(hash)
@@ -1847,7 +1458,7 @@ module Presto::Client::ModelVersions
     end
 
     class << SessionRepresentation =
-        Base.new(:query_id, :transaction_id, :client_transaction_support, :user, :principal, :source, :catalog, :schema, :path, :trace_token, :time_zone_key, :locale, :remote_user_address, :user_agent, :client_info, :client_tags, :client_capabilities, :resource_estimates, :start_time, :system_properties, :catalog_properties, :unprocessed_catalog_properties, :roles, :prepared_statements)
+        Base.new(:query_id, :transaction_id, :client_transaction_support, :user, :principal, :source, :catalog, :schema, :time_zone_key, :locale, :remote_user_address, :user_agent, :client_info, :start_time, :system_properties, :catalog_properties, :prepared_statements)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1862,21 +1473,14 @@ module Presto::Client::ModelVersions
           hash["source"],
           hash["catalog"],
           hash["schema"],
-          hash["path"] && SqlPath.decode(hash["path"]),
-          hash["traceToken"],
           hash["timeZoneKey"],
           hash["locale"],
           hash["remoteUserAddress"],
           hash["userAgent"],
           hash["clientInfo"],
-          hash["clientTags"],
-          hash["clientCapabilities"],
-          hash["resourceEstimates"] && ResourceEstimates.decode(hash["resourceEstimates"]),
           hash["startTime"],
           hash["systemProperties"],
           hash["catalogProperties"],
-          hash["unprocessedCatalogProperties"],
-          hash["roles"] && Hash[hash["roles"].to_a.map! {|k,v| [k, SelectedRole.decode(v)] }],
           hash["preparedStatements"],
         )
         obj
@@ -1904,7 +1508,7 @@ module Presto::Client::ModelVersions
     end
 
     class << SortNode =
-        Base.new(:id, :source, :ordering_scheme, :partial)
+        Base.new(:id, :source, :order_by, :orderings)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1913,15 +1517,15 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
-          hash["partial"],
+          hash["orderBy"],
+          hash["orderings"] && Hash[hash["orderings"].to_a.map! {|k,v| [k, v.downcase.to_sym] }],
         )
         obj
       end
     end
 
     class << Specification =
-        Base.new(:partition_by, :ordering_scheme)
+        Base.new(:partition_by, :order_by, :orderings)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -1929,7 +1533,8 @@ module Presto::Client::ModelVersions
         obj = allocate
         obj.send(:initialize_struct,
           hash["partitionBy"],
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
+          hash["orderBy"],
+          hash["orderings"] && Hash[hash["orderings"].to_a.map! {|k,v| [k, v.downcase.to_sym] }],
         )
         obj
       end
@@ -1949,57 +1554,8 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << SqlPath =
-        Base.new(:raw_path)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["rawPath"],
-        )
-        obj
-      end
-    end
-
-    class << StageExecutionDescriptor =
-        Base.new(:strategy, :grouped_execution_scan_nodes)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["strategy"] && hash["strategy"].downcase.to_sym,
-          hash["groupedExecutionScanNodes"],
-        )
-        obj
-      end
-    end
-
-    class << StageGcStatistics =
-        Base.new(:stage_id, :tasks, :full_gc_tasks, :min_full_gc_sec, :max_full_gc_sec, :total_full_gc_sec, :average_full_gc_sec)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["stageId"],
-          hash["tasks"],
-          hash["fullGcTasks"],
-          hash["minFullGcSec"],
-          hash["maxFullGcSec"],
-          hash["totalFullGcSec"],
-          hash["averageFullGcSec"],
-        )
-        obj
-      end
-    end
-
     class << StageInfo =
-        Base.new(:stage_id, :state, :self, :plan, :types, :stage_stats, :tasks, :sub_stages, :tables, :failure_cause)
+        Base.new(:stage_id, :state, :self, :plan, :types, :stage_stats, :tasks, :sub_stages, :failure_cause)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2014,7 +1570,6 @@ module Presto::Client::ModelVersions
           hash["stageStats"] && StageStats.decode(hash["stageStats"]),
           hash["tasks"] && hash["tasks"].map {|h| TaskInfo.decode(h) },
           hash["subStages"] && hash["subStages"].map {|h| StageInfo.decode(h) },
-          hash["tables"] && Hash[hash["tables"].to_a.map! {|k,v| [k, TableInfo.decode(v)] }],
           hash["failureCause"] && ExecutionFailureInfo.decode(hash["failureCause"]),
         )
         obj
@@ -2022,7 +1577,7 @@ module Presto::Client::ModelVersions
     end
 
     class << StageStats =
-        Base.new(:scheduling_complete, :get_split_distribution, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :blocked_drivers, :completed_drivers, :cumulative_user_memory, :user_memory_reservation, :revocable_memory_reservation, :total_memory_reservation, :peak_user_memory_reservation, :peak_revocable_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :physical_input_data_size, :physical_input_positions, :internal_network_input_data_size, :internal_network_input_positions, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :buffered_data_size, :output_data_size, :output_positions, :physical_written_data_size, :gc_info, :operator_summaries)
+        Base.new(:scheduling_complete, :get_split_distribution, :schedule_task_distribution, :add_split_distribution, :total_tasks, :running_tasks, :completed_tasks, :total_drivers, :queued_drivers, :running_drivers, :blocked_drivers, :completed_drivers, :cumulative_memory, :total_memory_reservation, :peak_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :buffered_data_size, :output_data_size, :output_positions, :operator_summaries)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2031,6 +1586,8 @@ module Presto::Client::ModelVersions
         obj.send(:initialize_struct,
           hash["schedulingComplete"],
           hash["getSplitDistribution"] && DistributionSnapshot.decode(hash["getSplitDistribution"]),
+          hash["scheduleTaskDistribution"] && DistributionSnapshot.decode(hash["scheduleTaskDistribution"]),
+          hash["addSplitDistribution"] && DistributionSnapshot.decode(hash["addSplitDistribution"]),
           hash["totalTasks"],
           hash["runningTasks"],
           hash["completedTasks"],
@@ -2039,21 +1596,15 @@ module Presto::Client::ModelVersions
           hash["runningDrivers"],
           hash["blockedDrivers"],
           hash["completedDrivers"],
-          hash["cumulativeUserMemory"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["cumulativeMemory"],
           hash["totalMemoryReservation"],
-          hash["peakUserMemoryReservation"],
-          hash["peakRevocableMemoryReservation"],
+          hash["peakMemoryReservation"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
+          hash["totalUserTime"],
           hash["totalBlockedTime"],
           hash["fullyBlocked"],
           hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
           hash["rawInputDataSize"],
           hash["rawInputPositions"],
           hash["processedInputDataSize"],
@@ -2061,8 +1612,6 @@ module Presto::Client::ModelVersions
           hash["bufferedDataSize"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
-          hash["gcInfo"] && StageGcStatistics.decode(hash["gcInfo"]),
           hash["operatorSummaries"] && hash["operatorSummaries"].map {|h| OperatorStats.decode(h) },
         )
         obj
@@ -2070,7 +1619,7 @@ module Presto::Client::ModelVersions
     end
 
     class << StatementStats =
-        Base.new(:state, :queued, :scheduled, :nodes, :total_splits, :queued_splits, :running_splits, :completed_splits, :cpu_time_millis, :wall_time_millis, :queued_time_millis, :elapsed_time_millis, :processed_rows, :processed_bytes, :peak_memory_bytes, :spilled_bytes, :root_stage)
+        Base.new(:state, :queued, :scheduled, :nodes, :total_splits, :queued_splits, :running_splits, :completed_splits, :user_time_millis, :cpu_time_millis, :wall_time_millis, :processed_rows, :processed_bytes, :root_stage)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2085,105 +1634,19 @@ module Presto::Client::ModelVersions
           hash["queuedSplits"],
           hash["runningSplits"],
           hash["completedSplits"],
+          hash["userTimeMillis"],
           hash["cpuTimeMillis"],
           hash["wallTimeMillis"],
-          hash["queuedTimeMillis"],
-          hash["elapsedTimeMillis"],
           hash["processedRows"],
           hash["processedBytes"],
-          hash["peakMemoryBytes"],
-          hash["spilledBytes"],
           hash["rootStage"] && ClientStageStats.decode(hash["rootStage"]),
         )
         obj
       end
     end
 
-    class << StatisticAggregations =
-        Base.new(:aggregations, :grouping_symbols)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["aggregations"] && Hash[hash["aggregations"].to_a.map! {|k,v| [k, Aggregation.decode(v)] }],
-          hash["groupingSymbols"],
-        )
-        obj
-      end
-    end
-
-    class << StatisticAggregationsDescriptor_Symbol =
-        Base.new(:grouping, :table_statistics, :column_statistics)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["grouping"],
-          hash["tableStatistics"] && Hash[hash["tableStatistics"].to_a.map! {|k,v| [k.downcase.to_sym, v] }],
-          hash["columnStatistics"] && Hash[hash["columnStatistics"].to_a.map! {|k,v| [ColumnStatisticMetadata.decode(k), v] }],
-        )
-        obj
-      end
-    end
-
-    class << StatisticsWriterNode =
-        Base.new(:id, :source, :target, :row_count_symbol, :row_count_enabled, :descriptor)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["id"],
-          hash["source"] && PlanNode.decode(hash["source"]),
-          hash["target"] && WriteStatisticsTarget.decode(hash["target"]),
-          hash["rowCountSymbol"],
-          hash["rowCountEnabled"],
-          hash["descriptor"] && StatisticAggregationsDescriptor_Symbol.decode(hash["descriptor"]),
-        )
-        obj
-      end
-    end
-
-    class << StatsAndCosts =
-        Base.new(:stats, :costs)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["stats"] && Hash[hash["stats"].to_a.map! {|k,v| [k, PlanNodeStatsEstimate.decode(v)] }],
-          hash["costs"] && Hash[hash["costs"].to_a.map! {|k,v| [k, PlanCostEstimate.decode(v)] }],
-        )
-        obj
-      end
-    end
-
-    class << SymbolStatsEstimate =
-        Base.new(:low_value, :high_value, :nulls_fraction, :average_row_size, :distinct_values_count)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["lowValue"],
-          hash["highValue"],
-          hash["nullsFraction"],
-          hash["averageRowSize"],
-          hash["distinctValuesCount"],
-        )
-        obj
-      end
-    end
-
     class << TableFinishInfo =
-        Base.new(:connector_output_metadata, :json_length_limit_exceeded, :statistics_wall_time, :statistics_cpu_time)
+        Base.new(:connector_output_metadata)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2191,16 +1654,13 @@ module Presto::Client::ModelVersions
         obj = allocate
         obj.send(:initialize_struct,
           hash["connectorOutputMetadata"],
-          hash["jsonLengthLimitExceeded"],
-          hash["statisticsWallTime"],
-          hash["statisticsCpuTime"],
         )
         obj
       end
     end
 
     class << TableFinishNode =
-        Base.new(:id, :source, :target, :row_count_symbol, :statistics_aggregation, :statistics_aggregation_descriptor)
+        Base.new(:id, :source, :target, :outputs)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2210,48 +1670,45 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["target"] && WriterTarget.decode(hash["target"]),
-          hash["rowCountSymbol"],
-          hash["statisticsAggregation"] && StatisticAggregations.decode(hash["statisticsAggregation"]),
-          hash["statisticsAggregationDescriptor"] && StatisticAggregationsDescriptor_Symbol.decode(hash["statisticsAggregationDescriptor"]),
+          hash["outputs"],
         )
         obj
       end
     end
 
     class << TableHandle =
-        Base.new(:catalog_name, :connector_handle, :transaction, :layout)
+        Base.new(:connector_id, :connector_handle)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["catalogName"],
+          hash["connectorId"],
           hash["connectorHandle"],
-          hash["transaction"],
-          hash["layout"],
         )
         obj
       end
     end
 
-    class << TableInfo =
-        Base.new(:table_name, :predicate)
+    class << TableLayoutHandle =
+        Base.new(:connector_id, :transaction_handle, :connector_handle)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
         end
         obj = allocate
         obj.send(:initialize_struct,
-          hash["tableName"],
-          hash["predicate"],
+          hash["connectorId"],
+          hash["transactionHandle"],
+          hash["connectorHandle"],
         )
         obj
       end
     end
 
     class << TableScanNode =
-        Base.new(:id, :table, :output_symbols, :assignments)
+        Base.new(:id, :table, :output_symbols, :assignments, :layout, :current_constraint, :original_constraint)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2262,30 +1719,16 @@ module Presto::Client::ModelVersions
           hash["table"] && TableHandle.decode(hash["table"]),
           hash["outputSymbols"],
           hash["assignments"],
-        )
-        obj
-      end
-    end
-
-    class << TableWriterInfo =
-        Base.new(:page_sink_peak_memory_usage, :statistics_wall_time, :statistics_cpu_time, :validation_cpu_time)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["pageSinkPeakMemoryUsage"],
-          hash["statisticsWallTime"],
-          hash["statisticsCpuTime"],
-          hash["validationCpuTime"],
+          hash["layout"] && TableLayoutHandle.decode(hash["layout"]),
+          hash["currentConstraint"],
+          hash["originalConstraint"],
         )
         obj
       end
     end
 
     class << TableWriterNode =
-        Base.new(:id, :source, :target, :row_count_symbol, :fragment_symbol, :columns, :column_names, :partitioning_scheme, :statistics_aggregation, :statistics_aggregation_descriptor)
+        Base.new(:id, :source, :target, :columns, :column_names, :outputs, :partitioning_scheme)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2295,20 +1738,17 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["target"] && WriterTarget.decode(hash["target"]),
-          hash["rowCountSymbol"],
-          hash["fragmentSymbol"],
           hash["columns"],
           hash["columnNames"],
+          hash["outputs"],
           hash["partitioningScheme"] && PartitioningScheme.decode(hash["partitioningScheme"]),
-          hash["statisticsAggregation"] && StatisticAggregations.decode(hash["statisticsAggregation"]),
-          hash["statisticsAggregationDescriptor"] && StatisticAggregationsDescriptor_Symbol.decode(hash["statisticsAggregationDescriptor"]),
         )
         obj
       end
     end
 
     class << TaskInfo =
-        Base.new(:task_status, :last_heartbeat, :output_buffers, :no_more_splits, :stats, :needs_plan)
+        Base.new(:task_status, :last_heartbeat, :output_buffers, :no_more_splits, :stats, :needs_plan, :complete)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2321,13 +1761,14 @@ module Presto::Client::ModelVersions
           hash["noMoreSplits"],
           hash["stats"] && TaskStats.decode(hash["stats"]),
           hash["needsPlan"],
+          hash["complete"],
         )
         obj
       end
     end
 
     class << TaskStats =
-        Base.new(:create_time, :first_start_time, :last_start_time, :last_end_time, :end_time, :elapsed_time, :queued_time, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :blocked_drivers, :completed_drivers, :cumulative_user_memory, :user_memory_reservation, :revocable_memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :physical_input_data_size, :physical_input_positions, :internal_network_input_data_size, :internal_network_input_positions, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :physical_written_data_size, :full_gc_count, :full_gc_time, :pipelines)
+        Base.new(:create_time, :first_start_time, :last_start_time, :last_end_time, :end_time, :elapsed_time, :queued_time, :total_drivers, :queued_drivers, :queued_partitioned_drivers, :running_drivers, :running_partitioned_drivers, :blocked_drivers, :completed_drivers, :cumulative_memory, :memory_reservation, :system_memory_reservation, :total_scheduled_time, :total_cpu_time, :total_user_time, :total_blocked_time, :fully_blocked, :blocked_reasons, :raw_input_data_size, :raw_input_positions, :processed_input_data_size, :processed_input_positions, :output_data_size, :output_positions, :pipelines)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2348,28 +1789,21 @@ module Presto::Client::ModelVersions
           hash["runningPartitionedDrivers"],
           hash["blockedDrivers"],
           hash["completedDrivers"],
-          hash["cumulativeUserMemory"],
-          hash["userMemoryReservation"],
-          hash["revocableMemoryReservation"],
+          hash["cumulativeMemory"],
+          hash["memoryReservation"],
           hash["systemMemoryReservation"],
           hash["totalScheduledTime"],
           hash["totalCpuTime"],
+          hash["totalUserTime"],
           hash["totalBlockedTime"],
           hash["fullyBlocked"],
           hash["blockedReasons"] && hash["blockedReasons"].map {|h| h.downcase.to_sym },
-          hash["physicalInputDataSize"],
-          hash["physicalInputPositions"],
-          hash["internalNetworkInputDataSize"],
-          hash["internalNetworkInputPositions"],
           hash["rawInputDataSize"],
           hash["rawInputPositions"],
           hash["processedInputDataSize"],
           hash["processedInputPositions"],
           hash["outputDataSize"],
           hash["outputPositions"],
-          hash["physicalWrittenDataSize"],
-          hash["fullGcCount"],
-          hash["fullGcTime"],
           hash["pipelines"] && hash["pipelines"].map {|h| PipelineStats.decode(h) },
         )
         obj
@@ -2377,7 +1811,7 @@ module Presto::Client::ModelVersions
     end
 
     class << TaskStatus =
-        Base.new(:task_id, :task_instance_id, :version, :state, :self, :node_id, :completed_driver_groups, :failures, :queued_partitioned_drivers, :running_partitioned_drivers, :output_buffer_overutilized, :physical_written_data_size, :memory_reservation, :system_memory_reservation, :revocable_memory_reservation, :full_gc_count, :full_gc_time)
+        Base.new(:task_id, :task_instance_id, :version, :state, :self, :failures, :queued_partitioned_drivers, :running_partitioned_drivers, :memory_reservation)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2389,25 +1823,17 @@ module Presto::Client::ModelVersions
           hash["version"],
           hash["state"] && hash["state"].downcase.to_sym,
           hash["self"],
-          hash["nodeId"],
-          hash["completedDriverGroups"] && hash["completedDriverGroups"].map {|h| Lifespan.new(h) },
           hash["failures"] && hash["failures"].map {|h| ExecutionFailureInfo.decode(h) },
           hash["queuedPartitionedDrivers"],
           hash["runningPartitionedDrivers"],
-          hash["outputBufferOverutilized"],
-          hash["physicalWrittenDataSize"],
           hash["memoryReservation"],
-          hash["systemMemoryReservation"],
-          hash["revocableMemoryReservation"],
-          hash["fullGcCount"],
-          hash["fullGcTime"],
         )
         obj
       end
     end
 
     class << TopNNode =
-        Base.new(:id, :source, :count, :ordering_scheme, :step)
+        Base.new(:id, :source, :count, :order_by, :orderings, :step)
       def decode(hash)
         unless hash.is_a?(Hash)
           raise TypeError, "Can't convert #{hash.class} to Hash"
@@ -2417,7 +1843,8 @@ module Presto::Client::ModelVersions
           hash["id"],
           hash["source"] && PlanNode.decode(hash["source"]),
           hash["count"],
-          hash["orderingScheme"] && OrderingScheme.decode(hash["orderingScheme"]),
+          hash["orderBy"],
+          hash["orderings"] && Hash[hash["orderings"].to_a.map! {|k,v| [k, v.downcase.to_sym] }],
           hash["step"] && hash["step"].downcase.to_sym,
         )
         obj
@@ -2512,50 +1939,6 @@ module Presto::Client::ModelVersions
       end
     end
 
-    class << Warning =
-        Base.new(:warning_code, :message)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["warningCode"] && Code.decode(hash["warningCode"]),
-          hash["message"],
-        )
-        obj
-      end
-    end
-
-    class << WarningCode =
-        Base.new(:code, :name)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["code"],
-          hash["name"],
-        )
-        obj
-      end
-    end
-
-    class << WindowInfo =
-        Base.new(:window_infos)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["windowInfos"] && hash["windowInfos"].map {|h| DriverWindowInfo.decode(h) },
-        )
-        obj
-      end
-    end
-
     class << WindowNode =
         Base.new(:id, :source, :specification, :window_functions, :hash_symbol, :pre_partitioned_inputs, :pre_sorted_order_prefix)
       def decode(hash)
@@ -2571,20 +1954,6 @@ module Presto::Client::ModelVersions
           hash["hashSymbol"],
           hash["prePartitionedInputs"],
           hash["preSortedOrderPrefix"],
-        )
-        obj
-      end
-    end
-
-    class << WriteStatisticsHandle =
-        Base.new(:handle)
-      def decode(hash)
-        unless hash.is_a?(Hash)
-          raise TypeError, "Can't convert #{hash.class} to Hash"
-        end
-        obj = allocate
-        obj.send(:initialize_struct,
-          hash["handle"] && AnalyzeTableHandle.decode(hash["handle"]),
         )
         obj
       end
