+++ spec/gzip_spec.rb
@@ -1,12 +1,12 @@
 require 'spec_helper'
 
-describe Presto::Client::Client do
+describe Trino::Client::Client do
   before(:all) do
     @spec_path = File.dirname(__FILE__)
     WebMock.disable!
-    @cluster = TinyPresto::Cluster.new('ghcr.io/trinodb/presto', '316')
+    @cluster = TinyPresto::Cluster.new()
     @container = @cluster.run
-    @client = Presto::Client.new(server: 'localhost:8080', catalog: 'tpch', user: 'test-user', schema: 'tiny', gzip: true, http_debug: true)
+    @client = Trino::Client.new(server: 'localhost:8080', catalog: 'tpch', user: 'test-user', schema: 'tiny', gzip: true, http_debug: true)
     loop do
       begin
         # Make sure to all workers are available.
